fix: import after move components
