+++ src/server/api/profile.ts
@@ -1,9 +1,10 @@
 import clerkClient from "@clerk/clerk-sdk-node"
-import type { PostAuthor, Profile, ProfileExtend } from "~/components/profilePage/types"
+
 import { filterClarkClientToAuthor, getFullName } from "~/utils/helpers"
 import { prisma } from "../db"
 import { type CacheSpecialKey, getCacheData, setCacheData } from "../cache"
 import { userFollowFollowedCount } from "./follow"
+import type { PostAuthor, Profile, ProfileExtend } from "~/features/profile"
 
 const MAX_CACHE_USER_LIFETIME_IN_SECONDS = 600
